Fixed High Latency Crash in LagCompensationComponent
Bug: From Lecture 205 - Projectile Score Request - When you shoot the host with high ping, with a server side rewind weapon, sometimes the engine would crash with this error:

Assertion failed: Pair != nullptr [File:E:\Unreal Engine\UE_5.1\Engine\Source\Runtime\Core\Public\Containers\Map.h] [Line: 643]

FindChecked() checks to see if the Pair is null, and if it is null, it triggers an assertion which causes the Engine to crash. “ The simplest way to think of assert is that whatever is "asserted" must be true, or the program will stop running.”
The key is sometimes null in ResetBoxes(), which I checked with an on screen message, checking if the key is contained in the TMap or not.

I just kept shooting at the host and eventually, it said false for the ResetBoxes() key.